chore : Enum, Struct 파일 분리
