+++ JuiceMaker/JuiceMaker/Model/Fruit.swift
@@ -1,0 +1,18 @@
+//
+//  Fruit.swift
+//  JuiceMaker
+//
+//  Created by 추니 & Davy.
+//
+
+import Foundation
+
+enum Fruit: CaseIterable {
+    case strawberry
+    case banana
+    case pineapple
+    case kiwi
+    case mango
+    
+    static let initialCount = 10
+}
