+++ JuiceMaker/JuiceMaker/Model/Juice.swift
@@ -1,0 +1,37 @@
+//
+//  Juice.swift
+//  JuiceMaker
+//
+//  Created by 추니 & Davy.
+//
+
+import Foundation
+
+enum Juice {
+    case strawberryJuice
+    case bananaJuice
+    case kiwiJuice
+    case pineappleJuice
+    case strawberryBananaJuice
+    case mangoJuice
+    case mangoKiwiJuice
+    
+    var name: String {
+        switch self {
+        case .strawberryJuice:
+            return "딸기쥬스"
+        case .bananaJuice:
+            return "바나나쥬스"
+        case .kiwiJuice:
+            return "키위쥬스"
+        case .pineappleJuice:
+            return "파인애플쥬스"
+        case .strawberryBananaJuice:
+            return "딸바쥬스"
+        case .mangoJuice:
+            return "망고쥬스"
+        case .mangoKiwiJuice:
+            return "망키쥬스"
+        }
+    }
+}
