+++ JuiceMaker/JuiceMaker/Model/JuiceResult.swift
@@ -1,0 +1,13 @@
+//
+//  JuiceResult.swift
+//  JuiceMaker
+//
+//  Created by 추니 & Davy.
+//
+
+import Foundation
+
+struct JuiceResult {
+    let isSuccess: Bool
+    let message: String?
+}
