+++ JuiceMaker/JuiceMaker/Model/Recipe.swift
@@ -1,0 +1,13 @@
+//
+//  Recipe.swift
+//  JuiceMaker
+//
+//  Created by 추니 & Davy.
+//
+
+import Foundation
+
+struct Recipe {
+    let fruit: Fruit
+    let amount: Int
+}
